+++ docs/numba.rst
@@ -1,0 +1,5 @@
+``fast_array_utils.numba``
+==========================
+
+.. automodule:: fast_array_utils.numba
+   :members:
