+++ pyproject.toml
@@ -50,8 +50,7 @@ doc = [
 ]
 test-min = [
   "coverage[toml]",
-  # include sparse for testing numba-less to_dense
-  "fast-array-utils[sparse,testing]",
+  "fast-array-utils[sparse,testing]", # include sparse for testing numba-less to_dense
   "pytest",
   "pytest-codspeed",
   "pytest-doctestplus",
