+++ src/fast_array_utils/stats/_mean_var.py
@@ -6,9 +6,8 @@ from typing import TYPE_CHECKING, no_type_check
 import numba
 import numpy as np
 
-from fast_array_utils._numba import njit
-
 from .. import types
+from ..numba import njit
 from ._power import power
