+++ typings/numba/core/config.pyi
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-from fast_array_utils._numba import Layer, LayerType
+from fast_array_utils.numba import TheadingCategory, ThreadingLayer
 
-THREADING_LAYER: Layer | LayerType
-THREADING_LAYER_PRIORITY: list[Layer]
+THREADING_LAYER: ThreadingLayer | TheadingCategory
+THREADING_LAYER_PRIORITY: list[ThreadingLayer]
